+++ include/boost/beast/_experimental/test/stream.hpp
@@ -32,6 +32,16 @@
 #include <mutex>
 #include <utility>
 
+#if ! BOOST_BEAST_DOXYGEN
+namespace boost {
+namespace asio {
+namespace ssl {
+template<typename> class stream;
+} // ssl
+} // asio
+} // boost
+#endif
+
 namespace boost {
 namespace beast {
 namespace test {
@@ -173,6 +183,27 @@ class stream
         boost::shared_ptr<state> const& in,
         std::unique_ptr<read_op_base>&& op,
         std::size_t buf_size);
+
+#if ! BOOST_BEAST_DOXYGEN
+    // boost::asio::ssl::stream needs these
+    // DEPRECATED
+    template<class>
+    friend class boost::asio::ssl::stream;
+    // DEPRECATED
+    using lowest_layer_type = stream;
+    // DEPRECATED
+    lowest_layer_type&
+    lowest_layer() noexcept
+    {
+        return *this;
+    }
+    // DEPRECATED
+    lowest_layer_type const&
+    lowest_layer() const noexcept
+    {
+        return *this;
+    }
+#endif
 
 public:
     using buffer_type = flat_buffer;